pyproject.toml
[tool.poetry]
name = "symptom-recognition"
version = "0.1.0"
description = "Thesis project on symptom recognition in medical texts"
authors = ["Igor Ivanov <gfn.ivanov@gmail.com>"]
repository = "https://github.com/gfnIvanov/symptom_recognition"
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
transformers = "^4.30.2"
torch = "^2.0.1"
pyyaml = "^6.0"
click = "^8.1.4"
dvc = "^3.5.1"
mypy = "^1.4.1"
types-click = "^7.1.8"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pre-commit = "^3.3.3"
autoflake = "^2.2.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.poetry.scripts]
use_gpt_native = "src.models.native.ruGPT_native:use_gpt_native"
use_bert_native = "src.models.native.bioBERT_native:use_bert_native"
parse_xml = "src.data.process:parse_xml"
depers = "src.data.process:depersonalization"
